streamlit_website/requirements.txt
streamlit==1.44.1
pandas==2.2.3
matplotlib==3.10.1
requests==2.32.3
python-dotenv==1.1.0